refactor guidelines
