+++ src/lib/pages/guidelines/update-guidelines-form.component.ts
@@ -1,12 +1,12 @@
 import {Component, OnInit} from '@angular/core';
-import {InteroperabilityRecord, Type} from '../../../domain/eic-model';
+import {InteroperabilityRecord, Type} from '../../domain/eic-model';
 import {FormBuilder} from '@angular/forms';
 import {ActivatedRoute, Router} from '@angular/router';
 import {GuidelinesFormComponent} from "./guidelines-form.component";
-import {AuthenticationService} from "../../../services/authentication.service";
-import {ServiceProviderService} from "../../../services/service-provider.service";
-import {ResourceService} from "../../../services/resource.service";
-import {GuidelinesService} from "../../../services/guidelines.service";
+import {AuthenticationService} from "../../services/authentication.service";
+import {ServiceProviderService} from "../../services/service-provider.service";
+import {ResourceService} from "../../services/resource.service";
+import {GuidelinesService} from "../../services/guidelines.service";
 
 
 @Component({
